[metrics]: Widen integer data types in kernel

+++ libs/ccc_cuda_ext/metrics.cuh
@@ -28,8 +28,8 @@ auto ari_core_host(const T *parts,
 // Used in the coef API
 template <typename T, typename R>
 auto ari_core_device(const py::array_t<T, py::array::c_style> &parts,
-                     const size_t n_features,
-                     const size_t n_parts,
-                     const size_t n_objs,
+                     const uint64_t n_features,
+                     const uint64_t n_parts,
+                     const uint64_t n_objs,
                      const uint64_t batch_start = 0,
                      const uint64_t batch_size = 0) -> std::unique_ptr<thrust::device_vector<R>>;
